requirements.txt
Flask==3.0.3
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
click==8.1.7
Werkzeug==3.0.4
reportlab
flask-login
psycopg2-binary